+++ docs/src/global/app.ts
@@ -25,3 +25,4 @@ import '@deckdeckgo/color';
 import '@deckdeckgo/social';
 import '@deckdeckgo/youtube';
 import '@deckdeckgo/drag-resize-rotate';
+import '@deckdeckgo/demo';