+++ studio/src/app/models/data/slide.tsx
@@ -8,13 +8,18 @@ export enum SlideTemplate {
   QRCODE = 'qrcode',
   CHART = 'chart',
   POLL = 'poll',
-  'ASPECT-RATIO' = 'aspect-ratio'
+  'ASPECT-RATIO' = 'aspect-ratio',
 }
 
 export enum SlideChartType {
   LINE = 'line',
   PIE = 'pie',
-  BAR = 'bar'
+  BAR = 'bar',
+}
+
+export enum SlideSplitType {
+  DEFAULT = 'default',
+  DEMO = 'demo',
 }
 
 export type SlideAttributesYAxisDomain = 'max' | 'extent';
@@ -29,7 +34,7 @@ export interface SlideAttributes {
   content?: string;
   customQRCode?: boolean;
 
-  type?: SlideChartType;
+  type?: SlideChartType | SlideSplitType;
   innerRadius?: number;
   animation?: boolean;
   datePattern?: string;