+++ studio/src/app/utils/editor/create-slides.utils.tsx
@@ -91,7 +91,7 @@ export class CreateSlidesUtils {
     });
   }
 
-  public static createSlideSplit(attributes: SlideAttributes = undefined): Promise<JSX.IntrinsicElements> {
+  static createSlideSplit(attributes: SlideAttributes = undefined): Promise<JSX.IntrinsicElements> {
     return new Promise<JSX.IntrinsicElements>((resolve) => {
       if (!document) {
         resolve();
@@ -336,6 +336,29 @@ export class CreateSlidesUtils {
       }
 
       const slide: JSX.IntrinsicElements = <deckgo-slide-aspect-ratio key={uuid()} grid={true} editable={true}></deckgo-slide-aspect-ratio>;
+
+      resolve(slide);
+    });
+  }
+
+  static createSlideDemo(src: string = undefined, mode: 'md' | 'ios'): Promise<JSX.IntrinsicElements> {
+    return new Promise<JSX.IntrinsicElements>((resolve) => {
+      if (!document) {
+        resolve();
+        return;
+      }
+
+      const start = <section slot="start"></section>;
+
+      const end = <deckgo-demo slot="end" src={src} mode={mode}></deckgo-demo>;
+
+      // @ts-ignore
+      // prettier-ignore
+      const slide: JSX.IntrinsicElements = (<deckgo-slide-split key={uuid()} type="demo">
+            {start}
+            {end}
+          </deckgo-slide-split>
+      );
 
       resolve(slide);
     });