+++ webcomponents/slides/split/src/components.d.ts
@@ -14,6 +14,7 @@ export namespace Components {
         "hideContent": () => Promise<void>;
         "lazyLoadContent": () => Promise<void>;
         "revealContent": () => Promise<void>;
+        "type": "demo" | "default";
         "vertical": boolean;
     }
 }
@@ -33,6 +34,7 @@ declare namespace LocalJSX {
         "customActions"?: boolean;
         "customBackground"?: boolean;
         "onSlideDidLoad"?: (event: CustomEvent<void>) => void;
+        "type"?: "demo" | "default";
         "vertical"?: boolean;
     }
     interface IntrinsicElements {