+++ webcomponents/slides/split/src/components/slide/deckdeckgo-slide-split.tsx
@@ -7,13 +7,13 @@ import {
   beforeSwipe,
   lazyLoadContent,
   hideAllRevealElements,
-  showAllRevealElements
+  showAllRevealElements,
 } from '@deckdeckgo/slide-utils';
 
 @Component({
   tag: 'deckgo-slide-split',
   styleUrl: 'deckdeckgo-slide-split.scss',
-  shadow: true
+  shadow: true,
 })
 export class DeckdeckgoSlideSplit implements DeckdeckgoSlide {
   @Element() el: HTMLElement;
@@ -24,6 +24,8 @@ export class DeckdeckgoSlideSplit implements DeckdeckgoSlide {
   @Prop({reflectToAttr: true}) customBackground: boolean = false;
 
   @Prop({reflectToAttr: true}) vertical: boolean = false;
+
+  @Prop({reflectToAttr: true}) type: 'demo' | 'default' = 'default';
 
   async componentDidLoad() {
     await hideLazyLoadImages(this.el);
@@ -43,7 +45,15 @@ export class DeckdeckgoSlideSplit implements DeckdeckgoSlide {
 
   @Method()
   lazyLoadContent(): Promise<void> {
-    return lazyLoadContent(this.el);
+    return new Promise<void>(async (resolve) => {
+      const promises = [];
+      promises.push(lazyLoadContent(this.el));
+      promises.push(this.resizeDemo());
+
+      await Promise.all(promises);
+
+      resolve();
+    });
   }
 
   @Method()
@@ -54,6 +64,18 @@ export class DeckdeckgoSlideSplit implements DeckdeckgoSlide {
   @Method()
   hideContent(): Promise<void> {
     return hideAllRevealElements(this.el);
+  }
+
+  private async resizeDemo() {
+    if (this.type !== 'demo') {
+      return;
+    }
+
+    const element: HTMLElement = this.el.querySelector(':scope > deckgo-demo');
+
+    if (element && typeof (element as any).updateIFrame === 'function') {
+      await (element as any).updateIFrame();
+    }
   }
 
   render() {