+++ webcomponents/slides/split/CHANGELOG.md
@@ -1,3 +1,27 @@
+<a name="1.1.2"></a>
+
+# 1.1.2 (2020-05-07)
+
+### Fix
+
+- init `demo` component size properly in starter kit
+
+<a name="1.1.1"></a>
+
+# 1.1.1 (2020-05-07)
+
+### Features
+
+- expose an attribute type use in studio to handle the new `demo` component
+
+<a name="1.1.0"></a>
+
+# 1.1.0 (2020-05-07)
+
+### Features
+
+- update `slide-utils` to support the new component `deckgo-demo`
+
 <a name="1.0.0"></a>
 
 # 1.0.0 (2020-03-19)